Fix coding style and organisation issues in manual enrollments
Update EnrollStudentsForm so that it uses the new API of
the UsersSearchSelectField.

Fix minor issues in the manual enrollment HTML templates.
Improve the wording of the translated strings.

Assistants can not see the manual enrollment button.

Add instructions to teachers about manual enrollments.

Remove one unused import from tests.
