+++ course/long_urls.py
@@ -23,8 +23,8 @@ urlpatterns = [
     url(EDIT_URL_PREFIX + r'participants/$',
         staff_views.ParticipantsView.as_view(),
         name="participants"),
-    url(EDIT_URL_PREFIX + r'enroll_students/$',
-        views.EnrollStudentsView.as_view(),
+    url(EDIT_URL_PREFIX + r'enroll-students/$',
+        staff_views.EnrollStudentsView.as_view(),
         name="enroll-students"),
     url(EDIT_URL_PREFIX + r'groups/$',
         staff_views.GroupsView.as_view(),
