+++ course/templates/course/staff/enroll_students.html
@@ -4,27 +4,31 @@
 {% load bootstrap %}
 {% load static %}
 
-{% block edittitle %}{% trans "Enroll students" %}{% endblock %}
+{% block scripts %}
+{{ block.super }}
+<script src="{% static 'js/ajax_search_select.js' %}"></script>
+{% endblock scripts %}
+{% block title %}{% trans "Enroll students" %} | {{ block.super }}{% endblock %}
 {% block view_tag %}participants,enroll-students{% endblock %}
 
 {% block breadcrumblist %}
 {{ block.super }}
 <li><a href="{{ instance|url:'participants' }}">{% trans "Participants" %}</a></li>
-<li class="active">{% trans "Add students" %}</li>
+<li class="active">{% trans "Enroll students" %}</li>
 {% endblock %}
 
 {% block columns %}
-<h3>{% trans "Add students to this course" %}</h3>
 <div class="col-md-12">
-
+  <p>
+  {% trans "Teacher may manually enroll students in the course here. Enrollment time and audience restrictions do not apply here. If the course has an enrollment exercise, the manually enrolled students are not required to submit anything." %}
+  </p>
   <form method="post" class="well form">
-      {% csrf_token %}
-      <legend>{% trans "Add students" %}</legend>
-      {{ form|bootstrap }}
-      <button type="submit" class="btn btn-primary">
-          {% trans "Save" %}
-      </button>
+    {% csrf_token %}
+    <legend>{% trans "Enroll students in the course" %}</legend>
+    {{ form|bootstrap }}
+    <button type="submit" class="btn btn-primary">
+      {% trans "Enroll students" %}
+    </button>
   </form>
-
 </div>
 {% endblock %}
