+++ course/templates/course/staff/participants.html
@@ -14,6 +14,14 @@
 {% block columns %}
 <div class="col-md-12">
 
+    {% if is_teacher %}
+    <p>
+      <a class="btn btn-primary btn-sm" href="{{ instance|url:'enroll-students' }}">
+        <span class="glyphicon glyphicon-user" aria-hidden="true"></span>
+          {% trans "Enroll students" %}
+      </a>
+    </p>
+    {% endif %}
     <p class="filter-users">
       <small>{% trans "Filter users" %}:</small>
       {% for tag in tags %}
@@ -23,12 +31,7 @@
       </button>
       {% endfor %}
     </p>
-    <p>
-    	<a class="btn btn-primary btn-xs" href="{{ instance|url:'enroll-students' }}">
-    		<span class="glyphicon glyphicon-user" aria-hidden="true"></span>
-    		{% trans "Enroll students" %}
-    	</a>
-    </p>
+
     <p>
         {% trans "Number of students" %} <strong id="participants-number"></strong>
         {% if is_teacher %}
